refactor: rename ConfigReader to ConfigElements
BREAKING CHANGE: "ConfigReader" namespace no longer available
