+++ TableCraft.Editor/ViewModels/ConfigAttributeUsageViewModel.cs
@@ -6,7 +6,7 @@
 #endregion
 
 using System.Windows.Input;
-using TableCraft.Core.ConfigReader;
+using TableCraft.Core.ConfigElements;
 using ReactiveUI;
 
 namespace TableCraft.Editor.ViewModels;
